fix ConsultFinished events silently failing to deserialize
serde_json does not support u128 when used with internally-tagged
enums (#[serde(tag = "type")]) + #[serde(flatten)]. The intermediate
Value representation cannot round-trip u128 values, causing all
ConsultFinished events to be silently dropped during deserialization.

Change duration_ms from u128 to u64 (still holds 584M years of
milliseconds). Add tests that verify all event types round-trip
through serde correctly.

Found by running the test-monitor harness against the TUI — consults
would start but never finish in the display.

+++ src/server.rs
@@ -101,7 +101,7 @@ impl ConsultServer {
                 consult_llm_mcp::monitoring::emit(
                     consult_llm_mcp::monitoring::MonitorEvent::ConsultFinished {
                         id: consultation_id,
-                        duration_ms: start_time.elapsed().as_millis(),
+                        duration_ms: start_time.elapsed().as_millis() as u64,
                         success: true,
                         error: None,
                     },
@@ -111,7 +111,7 @@ impl ConsultServer {
                 consult_llm_mcp::monitoring::emit(
                     consult_llm_mcp::monitoring::MonitorEvent::ConsultFinished {
                         id: consultation_id,
-                        duration_ms: start_time.elapsed().as_millis(),
+                        duration_ms: start_time.elapsed().as_millis() as u64,
                         success: false,
                         error: Some(e.to_string()),
                     },
